rubocop -A: 327 files inspected, 9796 offenses detected, 9125 offenses corrected

+++ lib/tasks/content.rake
@@ -1,19 +1,17 @@
+# frozen_string_literal: true
+
 namespace :content do
-  desc "Purges all expired and unused content in Concerto"
-  task :purge_unused_content => :environment do
+  desc 'Purges all expired and unused content in Concerto'
+  task purge_unused_content: :environment do
     Content.all.each do |c|
-      if c.is_orphan? && c.is_expired?
-        c.destroy
-      end
+      c.destroy if c.is_orphan? && c.is_expired?
     end
   end
-  
-  desc "Purges all expired content in Concerto"
-  task :purge_expired_content => :environment do
+
+  desc 'Purges all expired content in Concerto'
+  task purge_expired_content: :environment do
     Content.all.each do |c|
-      if c.is_expired?
-        c.destroy
-      end
+      c.destroy if c.is_expired?
     end
-  end  
+  end
 end
